Refactor MapBuilder and add offline mode
* Move most platform-specific stuff into `platform` module.
* Add offline mode for vector tiles.
* Remove unneeded `layer/tile_provider` module.
